requirements.txt
pandas
streamlit
python-docx
langchain
openai
tenacity